% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pdf.R
\name{extract_pdf}
\alias{extract_pdf}
\title{A .pdf extract Function}
\usage{
extract_pdf(filename)
}
\arguments{
\item{filename}{Filename of the .pdf file}
}
\description{
This function allows you to convert a .pdf file to a .txt file
}
\examples{
extract_pdf()
}
\keyword{.pdf}
