% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msa_confirm.R
\name{msa_confirm}
\alias{msa_confirm}
\title{An MSA Confirm Function}
\usage{
msa_confirm(data, scale)
}
\arguments{
\item{data}{.csv file with variables in columns and observations in rows}

\item{scale}{a character vector specifying the variables that should be included}
}
\description{
This function allows you to convert a .pdf file to a .txt file
}
\examples{
msa_confirm()
}
\keyword{msa}
