% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msa_explore.R
\name{msa_explore}
\alias{msa_explore}
\title{An MSA Explore Function}
\usage{
msa_explore(data)
}
\arguments{
\item{data}{.csv file with variables in columns and observations in rows}
}
\description{
This function allows you to run an exploratory MSA
}
\examples{
msa_explore()
}
\keyword{msa}
